+++ packages/plugins/src/plugins/streamflow/merkleFetcher.ts
@@ -1,13 +1,12 @@
 import { NetworkId } from '@sonarwatch/portfolio-core';
+import axios from 'axios';
+import BigNumber from 'bignumber.js';
 import { Cache } from '../../Cache';
 import { Fetcher, FetcherExecutor } from '../../Fetcher';
-import { platformId } from './constants';
-import { getClientSolana } from '../../utils/clients';
-import { getParsedMultipleAccountsInfo } from '../../utils/solana';
-import { claimStatusStruct } from './structs';
+import { platformId, streamflowApi } from './constants';
 import { MemoizedCache } from '../../utils/misc/MemoizedCache';
-import { getPdas } from './helpers';
-import { MerkleInfo } from './types';
+import { getPda } from './helpers';
+import { AirdropsResponse, MerkleInfo } from './types';
 import { ElementRegistry } from '../../utils/elementbuilder/ElementRegistry';
 
 const merkleMemo = new MemoizedCache<MerkleInfo[]>('merkles', {
@@ -16,45 +15,42 @@ const merkleMemo = new MemoizedCache<MerkleInfo[]>('merkles', {
 });
 
 const executor: FetcherExecutor = async (owner: string, cache: Cache) => {
-  const client = getClientSolana();
   const merkles = await merkleMemo.getItem(cache);
   if (!merkles) throw new Error('No active merkles found in cache');
 
-  const pdas = getPdas(
-    owner,
-    merkles.map((m) => m.address)
-  );
-
-  // This will only return accounts that exist
-  // meaning if the user didn't start to claim we won't but able to find his account
-  const claimAccounts = await getParsedMultipleAccountsInfo(
-    client,
-    claimStatusStruct,
-    pdas
+  const apiResponses = await Promise.all(
+    merkles.map((merkle) =>
+      axios
+        .get(`${streamflowApi}/airdrops/${merkle.address}/claimants/${owner}`)
+        .then((response) => response.data as AirdropsResponse)
+        .catch(() => undefined)
+    )
   );
 
   const registry = new ElementRegistry(NetworkId.solana, platformId);
-  for (let i = 0; i < claimAccounts.length; i += 1) {
-    const claim = claimAccounts[i];
-    if (!claim) continue;
+  for (let i = 0; i < apiResponses.length; i += 1) {
+    const allocationStatus = apiResponses[i];
+    if (!allocationStatus) continue;
 
     const merkle = merkles[i];
-    if (claim.closed) continue;
+    if (allocationStatus.chain !== 'SOLANA') continue;
 
     const element = registry.addElementMultiple({
       label: 'Airdrop',
       link: `https://app.streamflow.finance/airdrops/solana/mainnet/${merkle.address}`,
-      ref: claim.pubkey.toString(),
+      ref: getPda(owner, allocationStatus.distributorAddress),
       sourceRefs: [{ address: merkle.address, name: 'Distributor' }],
     });
     element.addAsset({
       address: merkle.mint,
-      amount: claim.unlockedAmount.minus(claim.lockedAmountWithdrawn),
+      amount: new BigNumber(allocationStatus.amountUnlocked).minus(
+        allocationStatus.amountClaimed
+      ),
       attributes: { isClaimable: true },
     });
     element.addAsset({
       address: merkle.mint,
-      amount: claim.lockedAmount.minus(claim.lockedAmountWithdrawn),
+      amount: new BigNumber(allocationStatus.amountLocked),
       attributes: { lockedUntil: -1 },
     });
   }